requirements.txt
numpy
pymongo
python-dotenv
certifi
ollama
dnspython
